DjangoInterface/requirements.txt
Django

django_extensions
djangorestframework
djangorestframework-jwt

channels
pathlib
asgi_redis
service_identity

Pillow
tensorflow
keras==2.1.5
pandas
numpy
h5py

boto3
tqdm